refactor: Controller 제거 리팩토링
