--- src/controller/ResultController.js
@@ -1,7 +1,0 @@
-import LottoResult from "../domain/LottoResult.js";
-
-export const ResultController = (winningLotto, lottoArray) => {
-  const lottoResult = new LottoResult(winningLotto, lottoArray);
-  const matchingCount = lottoResult.calculateResult();
-  return matchingCount;
-};
