+++ src/view/OutputView.js
@@ -10,8 +10,8 @@ const OutputView = {
   printLottoArray(lottoArray) {
     lottoArray.forEach((lotto) => OutputView.print(lotto.toString()));
   },
-  printMatchingCount(matchingCount) {
-    OutputView.print(SYSTEM_MESSAGE.WINNING_STATISTICS(matchingCount));
+  printMatchingResult(matchingResult) {
+    OutputView.print(SYSTEM_MESSAGE.WINNING_STATISTICS(matchingResult));
   },
 };
 export default OutputView;
